refactor: hook index.ts로 분리
