+++ src/hooks/useModal.tsx
@@ -6,7 +6,7 @@ import { useState } from 'react';
  * @param openModal 모달을 여는 함수입니다.
  * @param ModalComponent 모달 컴포넌트 입니다. 모달 안에  jsx 요소를 넣어 모달 컨텐츠로 쓸 수 있습니다.
  */
-function useModal() {
+export function useModal() {
     const [isOpen, setIsOpen] = useState<boolean>(false);
 
     const openModal = () => setIsOpen(true);
@@ -19,5 +19,3 @@ function useModal() {
         ModalComponent
     };
 }
-
-export default useModal;
